+++ CHANGES.rst
@@ -9,6 +9,7 @@
 * New render_with view decorator
 * New gfm module for GitHub Flavoured Markdown
 * load_models now supports "redirect" models.
+* Logging now looks for MAIL_DEFAULT_SENDER before DEFAULT_MAIL_SENDER
 
 0.3.11
 ------